Tests/LibWeb/Text/expected/HTML/form-add-dynamic-formdata.txt

formData.get('field1') === 'value1': true
formData.get('field2') === 'value2': true
